Try to guess what colour to set the SignColumn
Adds a new option, g:gitgutter_fix_bg.

If set to 1, it will try to steal the colour scheme of the user's
numberline (LineNr) if they are using it, otherwise will try to set
it based on &background. If set to 'light' or 'dark' it will try
to set it as if &background were set to that (even if it's not).
If it steals the highlighting from LineNr, it tries to figure
out if it seems light or dark.  It uses this to choose a brighter
or darker palette for the sign symbols.

If set to 0, nothing happens.  Default is currently 1, to make it more
likely that the thing "just works".