Remove unnecessary escaping.
While I can't say I fully understand escaping rules for Windows, I
understand them better now than I did when I copied that code from
another plugin and hoped for the best.

As far as I can tell, this escaping code was written to be used when
manually constructing a system command to execute – which isn't how this
plugin is written.  So this code shouldn't be here.