Better solution
I found that the my previous solution (ignoreing the sign at line 0)
cause annoying gutter resize if the removed first line is the only
modification of the file.

Thus I think placing the sign at line 1 instead of line 0 is the better
way while if the line 1 is modified, the sign would be overwritten.