Make sign line filter expression more specific
With previous expression, setting verbose>=15 
(which prints each line before it’s executed)
results in an exception because the “signs”
variable now includes the text of line 77, thus 
passing the filter. When the “name” variable 
assignment is then attempted, “components[2]”
contains no ‘=‘ signs and hence trying to access 
the second element of “split(components[2], ‘=‘)” 
results in an out-of-bounds exception (E684).