Initial support for hunks on staged mode
This includes initial, alpha support to "revert" staged changes. What is
really does is reset the hunk.

There are still problems. Since the staged mode hunk list is
non-existant/wrong, the positions are not corrected, so there's an
offset. If you correct that offset manually (relaxed current hunk
checks), the patch is applied correctly and the chunks are reset.

The priority now should be generating the correct hunks for the staged
mode.