Add neovim hunk preview floating window
If nvim is running and has floating window capabilities, display hunk
preview in a floating window instead of a preview windows. Then a user
can focus to that floating window with Ctrl-w to be able to scroll if
needed to. Moving the cursor in the current buffer closes the floating
window or :q inside the floating window.