[doc] metion the plugin vim-tmux-focus-events
By default tmux doesn't send focus-events. The result is that the
autocmd FocusGained isn't triggered. The docs already mention to set
`set -g focus-events on` in the tmux.conf However, gitgutter still
wouldn't update the signs when I committed the changes outside of vim.
To get it working I had to install the following plugin:
https://github.com/tmux-plugins/vim-tmux-focus-events
I think it would be great to mention this plugin in the docs so other
tmux-users are able to fix this issue quickly.