Revert GitGutterGetHunkSummary() to show hunks when inactive
In da77e46310e7e88fac8c837b3065341c3c944b87 the function changed to be
consistent with GitGutterGetHunks(), i.e. to show "no data" if called in
a buffer for which gitgutter is inactive.

This function is mostly used in status lines.  GitGutter is inactive
when a popup window is open.  The commit mentioned above meant that when
a completion window was open, the hunk summary in the status line
changed to show no data.  It changed back again once the completion
window was closed, but the overall effect was a little distracting.

This commit reverts the change.

Closes #771.